feat: add new Java examples for arithmetic operations, robot arm simulation, and triangle area calculation

+++ prakdaspro/ContohTipeData07.java
@@ -12,7 +12,7 @@ public class ContohTipeData07 {
     System.err.println("Jarak\t\t\t\t\t\t\t\t: " + jarak);
     System.err.println("Jumlah penduduk dalam satu dusun\t\t\t\t: " + jumlahPendudukDalamSetahun);
     System.err.println("Suhu\t\t\t\t\t\t\t\t: " + suhu);
-    System.err.println("Berat\t\t\t\t\t\t\t\t: " + berat);
+    System.err.println("Berat\t\t\t\t\t\t\t\t: " + (float) berat);
     System.err.println("Saldo\t\t\t\t\t\t\t\t: " + saldo);
     System.err.println("Angka desimal\t\t\t\t\t\t\t: " + angkaDesimal);
   }
